news brief generator/requirements.txt
groq>=0.4.0
python-dotenv>=0.19.0